Add training configuration and refactor training process in KModel; integrate TensorBoard logging and improve model training flow
